pyproject.toml
[project]
name = "langchain-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.4.0",
    "langchain>=1.1.0",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-groq>=1.1.0",
    "python-dotenv>=1.0.0",
    "wikipedia>=1.4.0",
]
